requirements.txt
flask
folium
qrcode
pillow
geopy
python-dateutil
matplotlib
numpy